+++ src/TimeUnit.php
@@ -4,15 +4,24 @@ declare(strict_types=1);
 
 namespace Bakame\Aide\Profiler;
 
+use Throwable;
 use ValueError;
 
 use function fmod;
 use function is_float;
 use function number_format;
+use function preg_match;
 use function round;
+use function strtolower;
 
 enum TimeUnit: int
 {
+    private const REGEXP_PATTERN = '/^
+        (?<number>\d+(?:\.\d+)?)
+        \s*
+        (?<unit>n|us|µs|ms|s|min|h)
+    $/ix';
+
     case Hour = 1_000 ** 3 * 3_600;
     case Minute = 1_000 ** 3 * 60;
     case Second = 1_000 ** 3;
@@ -22,24 +31,14 @@ enum TimeUnit: int
 
     public function fromNanoseconds(float|int $nanoseconds): float|int
     {
-        0 <= $nanoseconds || throw new ValueError('duration cannot be negative');
-        if (0.0 === $nanoseconds || 0 === $nanoseconds) {
-            return 0;
-        }
-
-        $converted = $nanoseconds / $this->value;
+        $converted = self::filterDuration($nanoseconds) / $this->value;
 
         return is_float($converted) && 0.0 === fmod($converted, 1.0) ? (int) $converted : $converted;
     }
 
     public function toNanoseconds(float|int $value): int
     {
-        0 <= $value || throw new ValueError('duration cannot be negative');
-        if (0.0 === $value || 0 === $value) {
-            return 0;
-        }
-
-        return (int) round($value * $this->value);
+        return (int) round(self::filterDuration($value) * $this->value);
     }
 
     /**
@@ -49,15 +48,18 @@ enum TimeUnit: int
      */
     public static function format(float|int $value, ?int $precision = null): string
     {
+        $value = self::filterDuration($value);
+        if (0 === $value) {
+            return '0 '.self::Nanosecond->suffix();
+        }
+
         foreach (self::cases() as $unit) {
             if ($value >= $unit->value) {
                 return $unit->formatFromNanoseconds($value, $precision).' '.$unit->suffix();
             }
         }
 
-        0 <= $value || throw new ValueError('duration cannot be negative');
-
-        return '0 '.self::Nanosecond->suffix();
+        return self::Nanosecond->formatFromNanoseconds($value, $precision).' '.self::Nanosecond->suffix();
     }
 
     private function formatFromNanoseconds(float|int $value, ?int $precision = null): string
@@ -79,5 +81,50 @@ enum TimeUnit: int
             self::Minute => 'min',
             self::Hour => 'h',
         };
+    }
+
+    /**
+     * @throws ValueError if bytes are not positive integer or zero
+     *
+     * @return float|int Zero or a positive numeric value.
+     */
+    private static function filterDuration(float|int $duration): float|int
+    {
+        0 <= $duration || throw new ValueError('duration cannot be negative');
+
+        if (0.0 === $duration || 0 === $duration) {
+            return 0;
+        }
+
+        return $duration;
+    }
+
+    /**
+     * @throws InvalidArgument If the parsing fails
+     */
+    public static function parse(string $value): int
+    {
+        1 === preg_match(self::REGEXP_PATTERN, $value, $matches) || throw new InvalidArgument('The value must be a valid time duration formatted string.');
+
+        $unit = match (strtolower($matches['unit'])) {
+            'n', 'ns' => self::Nanosecond,
+            'us', 'µs' => self::Microsecond,
+            'ms' => self::Millisecond,
+            's' => self::Second,
+            'min' => self::Minute,
+            'h' => self::Hour,
+            default => throw new InvalidArgument('Invalid or unsupported time unit.'),
+        };
+
+        return $unit->toNanoseconds((float) $matches['number']);
+    }
+
+    public static function tryParse(string $value): ?int
+    {
+        try {
+            return self::parse($value);
+        } catch (Throwable $exception) {
+            return null;
+        }
     }
 }